refactored mealLocationAdapter into its own package
